requirements.txt
setuptools
pybind11
torch
gym
numpy
scipy
psutil
py-cpuinfo
datasets
